feat: refactor map components and add directions panel with formatting utilities
