+++ roamance/src/utils/format.ts
@@ -1,0 +1,18 @@
+/**
+ * Format time in seconds to human-readable format
+ */
+export const formatTime = (seconds: number): string => {
+  if (seconds < 60) return `${seconds}s`;
+  if (seconds < 3600) return `${Math.floor(seconds / 60)}m`;
+  const hours = Math.floor(seconds / 3600);
+  const minutes = Math.floor((seconds % 3600) / 60);
+  return `${hours}h ${minutes > 0 ? ' ' + minutes + ' min' : ''}`;
+};
+
+/**
+ * Format distance in meters to human-readable format
+ */
+export const formatDistance = (meters: number): string => {
+  if (meters < 1000) return `${Math.round(meters)}m`;
+  return `${(meters / 1000).toFixed(1)}km`;
+};
